Read memory from the current pc memory space
Previously, it would read from memory space 0, which was not
always correct, and would error. This would lead to gdb assuming
that the value was all zeros, which prevents backtrace and
disassembly from working.

Now that we read from the correct memory space, bt and disass both
work as you would expect outside of EL3.