feat(UX): Allow un-prefixed components
Fast Models frequently have components follow a heirarchy
that starts with component.<fast-model-identifier>. This
is tedious to type out every time, and prevents the portability
of scripts across fast models.

This feature allows the use of the human-understandable component,
such as cluster0.cpu0 instead of component.FVP_Base_RevC_2xAEMvA.cluster0.cpu0

Further, this allows a script to be used with any model that has
a matching component instead of just the exact model specified in
the component path.