UX: print error descriptions
Previously, this would only print null unexpected. This was a bit
subtle, in that serde would atempt to deserialize the enum in
the order presented, and since the Response variant had a default
result to handle empty responses. This made the Response variant
match anything with an id, including errors.

Instead, reorder the enum to make error match first.