Feat: memory- info and translate subcommands
info looks up 'sideband information' including the start and
end addresses, the physical address that it maps to, and if
the core can execute code from that address (NX bit).

translate converts an address from one memory space to another,
listing all the addresses that the input address corresponds to.